+++ Backend/spring-boot-library/src/main/java/com/reactlibraryproject/springbootlibrary/Controller/CartItemController.java
@@ -2,6 +2,7 @@ package com.reactlibraryproject.springbootlibrary.Controller;
 
 import com.reactlibraryproject.springbootlibrary.ReponseModels.CurrentCartItemResponse;
 import com.reactlibraryproject.springbootlibrary.Service.BookService;
+import com.reactlibraryproject.springbootlibrary.Service.CartItemService;
 import com.reactlibraryproject.springbootlibrary.Utils.ExtractJWT;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -17,13 +18,13 @@ import java.util.List;
 @Tag(name="장바구니", description="장바구니 관련 API")
 public class CartItemController {
 
-    private BookService bookService;
+    private CartItemService cartItemService;
 
     @Operation(summary = "유저의 장바구니 목록")
     @GetMapping("/secure")
     public List<CurrentCartItemResponse> currentCart(@RequestHeader(value = "Authorization") String token) throws Exception {
         String userEmail = ExtractJWT.payloadJWTExtraction(token, "\"sub\"");
-        return bookService.currentCart(userEmail);
+        return cartItemService.currentCart(userEmail);
     }
 
     @Operation(summary = "장바구니에 책 추가")
@@ -31,7 +32,7 @@ public class CartItemController {
     public void addBookInCart(@RequestHeader(value = "Authorization") String token,
                               @RequestParam Long bookId) throws Exception {
         String userEmail = ExtractJWT.payloadJWTExtraction(token, "\"sub\"");
-        bookService.addBookInCart(userEmail, bookId);
+        cartItemService.addBookInCart(userEmail, bookId);
     }
 
     @Operation(summary = "장바구니에서 책 삭제")
@@ -39,7 +40,7 @@ public class CartItemController {
     public void deleteBookInCart(@RequestHeader(value = "Authorization") String token,
                                  @RequestParam Long bookId) throws Exception {
         String userEmail = ExtractJWT.payloadJWTExtraction(token, "\"sub\"");
-        bookService.deleteBookInCart(userEmail, bookId);
+        cartItemService.deleteBookInCart(userEmail, bookId);
     }
 
     @Operation(summary = "장바구니에서 책 수량 +")
@@ -47,7 +48,7 @@ public class CartItemController {
     public void increaseAmount(@RequestHeader(value = "Authorization") String token,
                                @RequestParam Long bookId) throws Exception {
         String userEmail = ExtractJWT.payloadJWTExtraction(token, "\"sub\"");
-        bookService.increaseAmount(userEmail, bookId);
+        cartItemService.increaseAmount(userEmail, bookId);
     }
 
     @Operation(summary = "장바구니에서 책 수량 -")
@@ -55,6 +56,6 @@ public class CartItemController {
     public void decreaseAmount(@RequestHeader(value = "Authorization") String token,
                                @RequestParam Long bookId) throws Exception {
         String userEmail = ExtractJWT.payloadJWTExtraction(token, "\"sub\"");
-        bookService.decreaseAmount(userEmail, bookId);
+        cartItemService.decreaseAmount(userEmail, bookId);
     }
 }
